pyproject.toml
[tool.poetry]
name = "developer-dashboard"
version = "0.1.0"
description = "A dashboard for developer to keep track of the developer velocity and the current burn down state of every task in the pipeline."
authors = ["aricma <adrian@aricma.com>"]
readme = "README.md"

packages = [
    { include = "server" },
    { include = "business_logic" },
    { include = "web_interface" },
]

[tool.poetry.dependencies]
python = "^3.8"
jira = { extras = ["cli"], version = "^3.5.2" }
humanize = "^4.9.0"
fastapi = "^0.105.0"
uvicorn = { extras = ["standard"], version = "^0.25.0" }
pyyaml = "^6.0.1"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
